refactor(artifact): mermaidTheme を components 配下から utils へ移動
components 配下に .ts を置かない方針に合わせて移動する。
Vue の reactivity を使わない設定値と純関数なので、CLAUDE.md の規約
(reactivity を使うものは composables/、使わないものは utils/) と
同じくアーティファクト系の純関数である utils/reactArtifactSrcdoc.ts の
前例に合わせて utils/ に置いた。
