+++ src/utils/mermaidTheme.ts
